+++ components/expansion-panel/stories/expansion-panel.stories.js
@@ -1,12 +1,13 @@
 import '@vonage/vwc-expansion-panel/vwc-expansion-panel.js';
 import '@vonage/vwc-textfield/vwc-textfield.js';
+import '@vonage/vwc-icon';
 import { html } from 'lit-element';
 import { spread } from '@open-wc/lit-helpers';
 import { argTypes } from './arg-types.js';
 
 export default {
-	title: 'Alpha/Components/Expansion Panel',
-	component: 'vwc-expansion-panel',
+	title: 'Alpha/Components/Accordion/Accordion Item',
+	component: 'vwc-accordion-item',
 	argTypes
 };
 
@@ -42,3 +43,20 @@ LeadingToggle.args = { header: 'Campaign Details', leadingToggle: '' };
 
 export const Icon = Template.bind({});
 Icon.args = { header: 'Campaign Details', icon: 'chat-solid' };
+
+const SlottedIconsTemplate = args => html`
+	<style>
+		vwc-icon {
+			--icon-size: 20px;
+		}
+	</style>
+  <vwc-expansion-panel ...=${spread(args)}>
+		<vwc-icon type="thumbs-up-line" slot="icon" size="small" connotation="cta"></vwc-icon>
+		<vwc-icon type="thumbs-down-line" slot="trailingIcon" connotation="info"></vwc-icon>
+		A campaign describes a specific use case and details of the messages you will be sending through it.
+		Such as: sample messages, subscriber opt-in/out, and the associated numbers you will be sending with these messages.
+  </vwc-expansion-panel>
+`;
+
+export const SlottedIcons = SlottedIconsTemplate.bind({});
+SlottedIcons.args = { header: 'Campaign Details' };