+++ components/side-drawer/src/vwc-side-drawer-base.ts
@@ -1,0 +1,71 @@
+import {
+	html, LitElement, TemplateResult, property,
+} from 'lit-element';
+import { nothing } from 'lit-html';
+import { ifDefined } from 'lit-html/directives/if-defined';
+import { classMap } from 'lit-html/directives/class-map';
+
+/**
+ * @slot header - The content of the header.
+ * @slot navigation - For vwc-list-item, vwc-list-expansion-panel, paragraph etc.
+ *
+ * @summary This is MyElement
+ *
+ */
+export class VWCSideDrawerBase extends LitElement {
+	/**
+	 * @prop alternate - [Applies scheme alternate region](../../common/scheme/readme.md)
+	 * accepts boolean value
+	 * @public
+	 * */
+	@property({ type: Boolean, reflect: true })
+	alternate = false;
+
+	/**
+	 * @prop hasTopBar - adds top bar to the side drawer
+	 * accepts boolean value
+	 * @public
+	 * */
+	@property({ type: Boolean, reflect: true })
+	hasTopBar?: boolean;
+
+	private renderTopBar(): TemplateResult {
+		return html`<div class="vvd-side-drawer--top-bar">
+			<slot name="top-bar"></slot>
+		</div>`;
+	}
+
+	/**
+	 * the html markup
+	 * @internal
+	 * */
+	protected render(): TemplateResult {
+		const topBar = this.hasTopBar
+			? this.renderTopBar()
+			: nothing;
+
+		const alternate = this.alternate
+			? 'vvd-scheme-alternate'
+			: undefined;
+
+		const classes = {
+			'vvd-side-drawer--alternate': this.alternate,
+			// 'vvd-side-drawer--modal': this.modal, // !@rinaok use with modal
+		};
+
+		return html`
+			<aside
+				part="${ifDefined(alternate)}"
+				class="side-drawer ${classMap(classes)}"
+			>
+
+				${topBar}
+
+				<div class="vvd-side-drawer--content">
+					<slot></slot>
+				</div>
+
+			</aside>
+		`;
+	}
+}