+++ components/snackbar/stories/snackbar.js
@@ -1,13 +1,11 @@
 import { COMPONENT_NAME } from '@vonage/vwc-snackbar';
 import '@vonage/vwc-button';
-import '@vonage/vwc-note';
-import '@vonage/vwc-radio';
 import { html } from 'lit-element';
 import { spread } from '@open-wc/lit-helpers';
 import { argTypes } from './arg-types.js';
 
 export default {
-	title: 'Alpha/Components/Snackbar',
+	title: 'Components/Snackbar',
 	component: COMPONENT_NAME,
 	argTypes
 };